+++ app/pages/Gifts/BuildGiftForm.view/BuildGiftForm.view.tsx
@@ -1,5 +1,5 @@
 import React, { useState } from 'react';
-import type { FC } from 'react';
+import type { ChangeEvent, FC } from 'react';
 
 import {
   Backdrop,
@@ -9,6 +9,7 @@ import {
   Fade,
   FormHelperText,
   FormLabel,
+  InputAdornment,
   LinearProgress,
   Slide,
   Modal,
@@ -21,21 +22,23 @@ import { useSnackbar } from 'notistack';
 import { useTranslation } from 'react-i18next';
 import * as Yup from 'yup';
 
-import { SubmitButton, MOBNumberFormat } from '../../../components';
+// import { AccountCard, SubmitButton, MOBNumberFormat } from '../../../../components';
+import AccountCard from '../../../components/AccountCard';
+import MOBNumberFormat from '../../../components/MOBNumberFormat';
+import SubmitButton from '../../../components/SubmitButton';
+import { MOBIcon } from '../../../components/icons';
 import useIsMountedRef from '../../../hooks/useIsMountedRef';
 import type { Theme } from '../../../theme';
 import type Account from '../../../types/Account';
-import { ConsumeGiftFormProps } from './ConsumeGiftForm.d';
+import { BuildGiftFormProps } from './BuildGiftForm';
 
 // CBB: Shouldn't have to use this hack to get around state issues
 const EMPTY_CONFIRMATION = {
+  feeConfirmation: null,
   giftCodeB58: '',
-  giftCodeStatus: '',
-  giftValue: 0,
+  totalValueConfirmation: null,
+  txProposal: null,
 };
-
-const CLAIMED_GIFT_ERROR =
-  '13 INTERNAL: transactions_manager.submit_tx_proposal: Connection(Operation { error: TransactionValidation(ContainsSpentKeyImage), total_delay: 0ns, tries: 1 })';
 
 const useStyles = makeStyles((theme: Theme) => ({
   button: {
@@ -66,38 +69,69 @@ const useStyles = makeStyles((theme: Theme) => ({
     alignItems: 'center',
     display: 'flex',
     justifyContent: 'center',
+    overflow: 'auto',
   },
   paper: {
     backgroundColor: theme.palette.background.paper,
     border: '2px solid #000',
     boxShadow: theme.shadows[5],
+    margin: '4rem',
+    maxHeight: '-webkit-fill-available',
     maxWidth: 800,
+    overflow: 'auto',
     padding: theme.spacing(2, 4, 3),
   },
   root: {},
 }));
 
+// TODO - ya, this definitely shouldn't live here
+const PICO_MOB_PRECISION = 12;
+
+const ensureMobStringPrecision = (mobString: string): string => {
+  const num = Number(mobString);
+  if (Number.isNaN(num)) {
+    throw new Error('mobString is NaN');
+  }
+
+  return num.toFixed(PICO_MOB_PRECISION);
+};
+
+// TODO - ya, this definitely shouldn't live here
+const convertMobStringToPicoMobString = (mobString: string): string =>
+  ensureMobStringPrecision(mobString).replace('.', '');
+
 // TODO -- right now, we can show a progress bar for the sending modal
 // But, it would be nice to have a counter that parses up to, say, 10 seconds, before
 // warning that it's taking a bit long...
 // TODO -- we may want to refactor out the modals and feed them props just to keep
-// this component managable.
-const ConsumeGiftForm: FC<ConsumeGiftFormProps> = ({
-  checkGiftCodeStatus,
-  claimGiftCode,
+// this component manageable.
+const BuildGiftForm: FC<BuildGiftFormProps> = ({
+  buildGiftCode,
+  existingPin,
+  isSyncedBuffered,
+  pinThresholdPmob,
   selectedAccount,
-}: ConsumeGiftFormProps) => {
+  submitGiftCode,
+}: BuildGiftFormProps) => {
   const classes = useStyles();
   const [confirmation, setConfirmation] = useState(EMPTY_CONFIRMATION);
   const [showModal, setShowModal] = useState(false);
+  const [showCode, setShowCode] = useState(false);
   const [isAwaitingConformation, setIsAwaitingConformation] = useState(false);
   const [submittingConfimedGift, setSubmittingConfirmedGift] = useState(false);
+  const [slideExitSpeed, setSlideExitSpeed] = useState(0);
   const { enqueueSnackbar } = useSnackbar();
   const isMountedRef = useIsMountedRef();
-  const { t } = useTranslation('ConsumeGiftForm');
 
+  const { t } = useTranslation('BuildGiftForm');
+
+  const networkBlockIndexBigInt = BigInt(selectedAccount.balanceStatus.networkBlockIndex);
+  const accountBlockIndexBigInt = BigInt(selectedAccount.balanceStatus.accountBlockIndex);
+
+  const isSynced = isSyncedBuffered(networkBlockIndexBigInt, accountBlockIndexBigInt);
+
+  // TODO - consider adding minimum gift ~ 1 MOB
   // We'll use this array in prep for future patterns with multiple accounts
-  // TODO - fix the type for Account
   const mockMultipleAccounts: Array<Account> = [
     {
       b58Code: selectedAccount.account.mainAddress,
@@ -106,7 +140,14 @@ const ConsumeGiftForm: FC<ConsumeGiftFormProps> = ({
     },
   ];
 
+  const handleShowCode = () => {
+    setSlideExitSpeed(1000);
+    setShowCode(true);
+  };
+
   const handleClose = () => {
+    setSlideExitSpeed(0);
+    setShowCode(false);
     setShowModal(false);
     setIsAwaitingConformation(false);
     setConfirmation(EMPTY_CONFIRMATION);
@@ -119,25 +160,28 @@ const ConsumeGiftForm: FC<ConsumeGiftFormProps> = ({
     setSubmittingConfirmedGift(true);
     setShowModal(false);
     try {
-      await claimGiftCode({
-        accountId: selectedAccount.account.accountId,
-        giftCodeB58: confirmation.giftCodeB58,
-      });
+      if (confirmation.txProposal === null || confirmation.txProposal === undefined) {
+        throw new Error(t('confirmationNotFound'));
+      }
 
+      await submitGiftCode({
+        fromAccountId: selectedAccount.account.accountId,
+        giftCodeB58: confirmation.giftCodeB58,
+        txProposal: confirmation.txProposal,
+      });
       if (isMountedRef.current) {
         setStatus({ success: true });
         setSubmittingConfirmedGift(false);
         setIsAwaitingConformation(false);
         setConfirmation(EMPTY_CONFIRMATION);
-        enqueueSnackbar(t('confirmation'), {
+        enqueueSnackbar(t('giftCreated'), {
           variant: 'success',
         });
       }
     } catch (err) {
       if (isMountedRef.current) {
-        const santitizedError = err.message === CLAIMED_GIFT_ERROR ? t('giftClaimed') : err.message;
         setStatus({ success: false });
-        setErrors({ submit: santitizedError });
+        setErrors({ submit: err.message });
         setSubmittingConfirmedGift(false);
         setIsAwaitingConformation(false);
         setConfirmation(EMPTY_CONFIRMATION);
@@ -148,7 +192,6 @@ const ConsumeGiftForm: FC<ConsumeGiftFormProps> = ({
     }
   };
 
-  // TODO - not sure what this is about. Should delete after confirming it's useless
   // const createAccountLabel = (account: Account) => {
   //   const name =
   //     account.name && account.name.length > 0 ? `${account.name}: ` : `${t('unnamed')}: `;
@@ -168,8 +211,8 @@ const ConsumeGiftForm: FC<ConsumeGiftFormProps> = ({
   //   );
   // };
 
-  // TODO Reintroduce with multiple accounts
-  // const renderSenderPublicAdddressOptions = (accounts: Account[], isSubmitting: boolean) => (
+  // TODO - reintroduce with multiple accounts
+  // const renderSenderPublicAddressOptions = (accounts: Account[], isSubmitting: boolean) => (
   //   <Box pt={2}>
   //     <FormLabel className={classes.form} component="legend">
   //       <Typography color="primary">{t('select')}</Typography>
@@ -197,50 +240,67 @@ const ConsumeGiftForm: FC<ConsumeGiftFormProps> = ({
   //   </Box>
   // );
 
+  const validateAmount = (selectedBalance: bigint, fee: bigint) => (valueString: string) => {
+    let error;
+    const valueAsPicoMob = BigInt(valueString.replace('.', ''));
+    if (valueAsPicoMob + fee + fee > selectedBalance) {
+      // TODO - probably want to replace this before launch
+      error = t('errorFee');
+    }
+    return error;
+  };
+
+  // We'll use this to auto-select all text when focused. This is a better user
+  // experience than having to click the left-most area to start typing (else)
+  // having to spam backspace.
+  const handleSelect = (event: ChangeEvent<HTMLInputElement>) => {
+    event.target.select();
+  };
+
   return (
     <Formik
       initialValues={{
-        giftCodeB58: '', // mobs
+        feeAmount: '0.010000000000', // TODO we need to pull this from constants
+        mobValue: '0', // mobs
+        pin: '',
         senderPublicAddress: mockMultipleAccounts[0].b58Code,
         submit: null,
       }}
       validationSchema={Yup.object().shape({
-        giftCodeB58: Yup.string().required(t('giftB58Validation')),
+        mobValue: Yup.number()
+          .positive(t('positiveValidation'))
+          .required(t('positiveValidationRequired')),
       })}
       onSubmit={async (values, { setErrors, setStatus, setSubmitting, resetForm }) => {
+        // On submit, let's build the TxProposal.
+        // On success, we set the TxProposal in state.
+        // That triggers the confirmation modal.
+        // In the modal, the user has to confirm that they have written the
+        // code down (similiar to create new account modal).
+        // After they do, they can "Create Gift"
         try {
           setIsAwaitingConformation(true);
-          const result = await checkGiftCodeStatus({ giftCodeB58: values.giftCodeB58 });
+
+          const adjustedValue = Number(values.mobValue) + 0.01;
+
+          const result = await buildGiftCode({
+            accountId: selectedAccount.account.accountId,
+            valuePmob: convertMobStringToPicoMobString(String(adjustedValue)),
+          });
           if (result === null || result === undefined) {
-            throw new Error(t('giftB58Error'));
+            throw new Error(t('errorBuild'));
           }
 
-          const { giftCodeStatus, giftCodeValue } = result;
+          const { feeConfirmation, giftCodeB58, totalValueConfirmation, txProposal } = result;
 
           setConfirmation({
-            giftCodeB58: values.giftCodeB58,
-            giftCodeStatus,
-            giftValue: giftCodeValue,
+            feeConfirmation,
+            giftCodeB58,
+            totalValueConfirmation,
+            txProposal,
           });
 
-          if (giftCodeStatus === 'GiftCodeAvailable') {
-            setShowModal(true);
-          } else {
-            setStatus({ success: false });
-            setSubmitting(false);
-            setIsAwaitingConformation(false);
-            if (giftCodeStatus === 'GiftCodeSubmittedPending') {
-              enqueueSnackbar(t('giftB58Error'), {
-                variant: 'warning',
-              });
-            }
-
-            if (giftCodeStatus === 'GiftCodeClaimed') {
-              enqueueSnackbar(t('giftClaimed'), {
-                variant: 'warning',
-              });
-            }
-          }
+          setShowModal(true);
 
           if (isMountedRef.current) {
             setSubmitting(false);
@@ -256,24 +316,35 @@ const ConsumeGiftForm: FC<ConsumeGiftFormProps> = ({
         }
       }}
     >
-      {({ errors, isSubmitting, dirty, isValid, submitForm, setErrors, setStatus, values }) => {
+      {({ errors, isSubmitting, isValid, dirty, submitForm, values, setErrors, setStatus }) => {
         const selectedBalance =
           // TODO -- this is fine. we'll gut it anyway once we add multiple accounts
           // eslint-disable-next-line
           // @ts-ignore
-          mockMultipleAccounts.find((account) => account.b58Code === values.senderPublicAddress)
-            .balance;
-        let increasedBalance;
-        let totalSent;
+          BigInt(
+            mockMultipleAccounts.find((account) => account.b58Code === values.senderPublicAddress)
+              .balance
+          );
 
-        if (confirmation?.giftValue) {
-          increasedBalance = Number(selectedBalance) + confirmation?.giftValue - 10000000000;
-          totalSent = confirmation?.giftValue;
+        let isPinRequiredForTransaction = false;
+        if (confirmation.totalValueConfirmation) {
+          isPinRequiredForTransaction =
+            confirmation?.totalValueConfirmation + confirmation?.feeConfirmation >=
+            BigInt(pinThresholdPmob);
+        }
+
+        let remainingBalance;
+        let totalSent;
+        if (confirmation?.totalValueConfirmation && confirmation?.feeConfirmation) {
+          remainingBalance =
+            selectedBalance -
+            (confirmation?.totalValueConfirmation + confirmation?.feeConfirmation);
+          totalSent = confirmation?.totalValueConfirmation + confirmation?.feeConfirmation;
         }
 
         return (
           <Form>
-            {/* {renderSenderPublicAdddressOptions(mockMultipleAccounts, isSubmitting)} */}
+            {/* {renderSenderPublicAddressOptions(mockMultipleAccounts, isSubmitting)} */}
             <Box pt={4}>
               <FormLabel component="legend">
                 <Typography color="primary">{t('giftDetails')}</Typography>
@@ -281,11 +352,24 @@ const ConsumeGiftForm: FC<ConsumeGiftFormProps> = ({
               <Field
                 component={TextField}
                 fullWidth
-                label={t('giftCode')}
+                label={t('mobLabel')}
                 margin="normal"
-                name="giftCodeB58"
-                id="giftCodeB58"
+                name="mobValue"
+                id="mobValue"
                 type="text"
+                onFocus={handleSelect}
+                validate={validateAmount(
+                  selectedBalance,
+                  BigInt(values.feeAmount * 1_000_000_000_000)
+                )}
+                InputProps={{
+                  inputComponent: MOBNumberFormat,
+                  startAdornment: (
+                    <InputAdornment position="start">
+                      <MOBIcon height={20} width={20} />
+                    </InputAdornment>
+                  ),
+                }}
               />
             </Box>
             {errors.submit && (
@@ -293,12 +377,17 @@ const ConsumeGiftForm: FC<ConsumeGiftFormProps> = ({
                 <FormHelperText error>{errors.submit}</FormHelperText>
               </Box>
             )}
+            {!isSynced && (
+              <Box mt={3}>
+                <FormHelperText error>{t('errorSyncBeforeSending')}</FormHelperText>
+              </Box>
+            )}
             <SubmitButton
-              disabled={!dirty || !isValid || isSubmitting}
+              disabled={!dirty || !isSynced || !isValid || isSubmitting}
               onClick={submitForm}
               isSubmitting={isAwaitingConformation || isSubmitting}
             >
-              {t('openGift')}
+              {isSynced ? t('createGift') : `${t('walletSyncing')}...`}
             </SubmitButton>
             <Modal
               aria-labelledby="transition-modal-title"
@@ -314,12 +403,15 @@ const ConsumeGiftForm: FC<ConsumeGiftFormProps> = ({
               disableAutoFocus
               disableEnforceFocus
             >
-              <Slide in={showModal} timeout={{ enter: 0, exit: 0 }}>
+              <Slide in={showModal} timeout={{ enter: 0, exit: slideExitSpeed }}>
                 <Container className={classes.paper}>
-                  <Typography color="textPrimary" variant="h1" id="transition-modal-title">
-                    {t('giftConfirmation')}
-                  </Typography>
-                  <Box py={2} />
+                  <Box py={2}>
+                    <Typography variant="h1" color="textPrimary">
+                      {t('giftConfirmation')}
+                    </Typography>
+                    {/* <Typography variant="p" color="textPrimary"> */}
+                    <Typography color="textPrimary">{t('giftConfirmationDescription')}:</Typography>
+                  </Box>
                   <Box display="flex" justifyContent="space-between">
                     <Typography color="textPrimary">{t('accountBalance')}:</Typography>
                     <Typography color="textPrimary">
@@ -335,12 +427,14 @@ const ConsumeGiftForm: FC<ConsumeGiftFormProps> = ({
                     <Typography color="textPrimary">---</Typography>
                   </Box>
                   <Box display="flex" justifyContent="space-between">
-                    <Typography color="textPrimary">{t('total')}:</Typography>
-                    <Typography color="textPrimary">
+                    <Typography color="primary">{t('giftValue')}:</Typography>
+                    <Typography color="primary">
                       <MOBNumberFormat
                         suffix=" MOB"
                         valueUnit="pMOB"
-                        value={totalSent?.toString() as string}
+                        value={(
+                          confirmation?.totalValueConfirmation - confirmation?.feeConfirmation
+                        ).toString()}
                       />
                     </Typography>
                   </Box>
@@ -350,36 +444,68 @@ const ConsumeGiftForm: FC<ConsumeGiftFormProps> = ({
                       <MOBNumberFormat
                         suffix=" MOB"
                         valueUnit="pMOB"
-                        value={(10000000000).toString()}
+                        value={(
+                          confirmation?.feeConfirmation + confirmation?.feeConfirmation
+                        ).toString()}
                       />
                     </Typography>
                   </Box>
                   <Box display="flex" justifyContent="space-between">
-                    <Typography color="primary">{t('giftValue')}:</Typography>
-                    <Typography color="primary">
+                    <Typography color="textPrimary">{t('total')}:</Typography>
+                    <Typography color="textPrimary">
                       <MOBNumberFormat
                         suffix=" MOB"
                         valueUnit="pMOB"
-                        value={(confirmation?.giftValue - 10000000000).toString()}
+                        value={totalSent?.toString()}
                       />
                     </Typography>
                   </Box>
                   <Box display="flex" justifyContent="space-between">
-                    <Typography>---</Typography>
-                    <Typography>---</Typography>
+                    <Typography color="textPrimary">---</Typography>
+                    <Typography color="textPrimary">---</Typography>
                   </Box>
                   <Box display="flex" justifyContent="space-between">
-                    <Typography color="primary">{t('newBalance')}:</Typography>
+                    <Typography color="primary">{t('remaining')}:</Typography>
                     <Typography color="primary">
                       <MOBNumberFormat
                         suffix=" MOB"
                         valueUnit="pMOB"
-                        value={increasedBalance?.toString() as string}
+                        value={remainingBalance?.toString()}
                       />
                     </Typography>
                   </Box>
                   <Box py={1} />
-                  <Box display="flex" justifyContent="space-between">
+                  {/* TODO - after multiple accounts, we should actually store these gift codes. please check jira for full explation */}
+                  <Typography variant="body2" color="textPrimary">
+                    {t('mobWillBeSent')}
+                  </Typography>
+                  <Box py={1} />
+                  {showCode ? (
+                    <AccountCard
+                      isGift
+                      account={{
+                        b58Code: confirmation?.giftCodeB58,
+                        name: t('pending'),
+                      }}
+                    />
+                  ) : (
+                    <Box display="flex" justifyContent="center" py={27}>
+                      <Button color="secondary" size="large" onClick={handleShowCode}>
+                        {t('showCode')}
+                      </Button>
+                    </Box>
+                  )}
+                  {isPinRequiredForTransaction && (
+                    <Field
+                      component={TextField}
+                      fullWidth
+                      label={t('enterPin')}
+                      margin="normal"
+                      name="pin"
+                      type="password"
+                    />
+                  )}
+                  <Box display="flex" justifyContent="space-between" style={{ padding: '1rem' }}>
                     <Button
                       className={classes.button}
                       color="secondary"
@@ -394,11 +520,14 @@ const ConsumeGiftForm: FC<ConsumeGiftFormProps> = ({
                     <Button
                       className={classes.button}
                       color="secondary"
+                      disabled={
+                        !showCode || (isPinRequiredForTransaction && values.pin !== existingPin)
+                      }
                       fullWidth
                       onClick={handleConfirm(setErrors, setStatus)}
                       variant="contained"
                     >
-                      {t('claimGift')}
+                      {showCode ? t('confirmGift') : t('secureCode')}
                     </Button>
                   </Box>
                 </Container>
@@ -429,5 +558,5 @@ const ConsumeGiftForm: FC<ConsumeGiftFormProps> = ({
   );
 };
 
-export default ConsumeGiftForm;
-export { ConsumeGiftForm };
+export default BuildGiftForm;
+export { BuildGiftForm };
