+++ app/pages/Gifts/BuildGiftForm.view/index.ts
@@ -1,4 +1,4 @@
-import { BuildGiftForm } from './BuildGiftForm';
+import { BuildGiftForm } from './BuildGiftForm.view';
 
 export default BuildGiftForm;
 export { BuildGiftForm };
