+++ app/pages/Gifts/ConsumeGiftForm.view/ConsumeGiftForm.view.tsx
@@ -1,5 +1,5 @@
 import React, { useState } from 'react';
-import type { ChangeEvent, FC } from 'react';
+import type { FC } from 'react';
 
 import {
   Backdrop,
@@ -9,7 +9,6 @@ import {
   Fade,
   FormHelperText,
   FormLabel,
-  InputAdornment,
   LinearProgress,
   Slide,
   Modal,
@@ -22,23 +21,21 @@ import { useSnackbar } from 'notistack';
 import { useTranslation } from 'react-i18next';
 import * as Yup from 'yup';
 
-// import { AccountCard, SubmitButton, MOBNumberFormat } from '../../../../components';
-import AccountCard from '../../../components/AccountCard';
-import MOBNumberFormat from '../../../components/MOBNumberFormat';
-import SubmitButton from '../../../components/SubmitButton';
-import { MOBIcon } from '../../../components/icons';
+import { SubmitButton, MOBNumberFormat } from '../../../components';
 import useIsMountedRef from '../../../hooks/useIsMountedRef';
 import type { Theme } from '../../../theme';
 import type Account from '../../../types/Account';
-import { BuildGiftFormProps } from './BuildGiftForm.d';
+import { ConsumeGiftFormProps } from './ConsumeGiftForm';
 
 // CBB: Shouldn't have to use this hack to get around state issues
 const EMPTY_CONFIRMATION = {
-  feeConfirmation: null,
   giftCodeB58: '',
-  totalValueConfirmation: null,
-  txProposal: null,
+  giftCodeStatus: '',
+  giftValue: 0,
 };
+
+const CLAIMED_GIFT_ERROR =
+  '13 INTERNAL: transactions_manager.submit_tx_proposal: Connection(Operation { error: TransactionValidation(ContainsSpentKeyImage), total_delay: 0ns, tries: 1 })';
 
 const useStyles = makeStyles((theme: Theme) => ({
   button: {
@@ -69,69 +66,38 @@ const useStyles = makeStyles((theme: Theme) => ({
     alignItems: 'center',
     display: 'flex',
     justifyContent: 'center',
-    overflow: 'auto',
   },
   paper: {
     backgroundColor: theme.palette.background.paper,
     border: '2px solid #000',
     boxShadow: theme.shadows[5],
-    margin: '4rem',
-    maxHeight: '-webkit-fill-available',
     maxWidth: 800,
-    overflow: 'auto',
     padding: theme.spacing(2, 4, 3),
   },
   root: {},
 }));
 
-// TODO - ya, this definitely shouldn't live here
-const PICO_MOB_PRECISION = 12;
-
-const ensureMobStringPrecision = (mobString: string): string => {
-  const num = Number(mobString);
-  if (Number.isNaN(num)) {
-    throw new Error('mobString is NaN');
-  }
-
-  return num.toFixed(PICO_MOB_PRECISION);
-};
-
-// TODO - ya, this definitely shouldn't live here
-const convertMobStringToPicoMobString = (mobString: string): string =>
-  ensureMobStringPrecision(mobString).replace('.', '');
-
 // TODO -- right now, we can show a progress bar for the sending modal
 // But, it would be nice to have a counter that parses up to, say, 10 seconds, before
 // warning that it's taking a bit long...
 // TODO -- we may want to refactor out the modals and feed them props just to keep
-// this component manageable.
-const BuildGiftForm: FC<BuildGiftFormProps> = ({
-  buildGiftCode,
-  existingPin,
-  isSyncedBuffered,
-  pinThresholdPmob,
+// this component managable.
+const ConsumeGiftForm: FC<ConsumeGiftFormProps> = ({
+  checkGiftCodeStatus,
+  claimGiftCode,
   selectedAccount,
-  submitGiftCode,
-}: BuildGiftFormProps) => {
+}: ConsumeGiftFormProps) => {
   const classes = useStyles();
   const [confirmation, setConfirmation] = useState(EMPTY_CONFIRMATION);
   const [showModal, setShowModal] = useState(false);
-  const [showCode, setShowCode] = useState(false);
   const [isAwaitingConformation, setIsAwaitingConformation] = useState(false);
   const [submittingConfimedGift, setSubmittingConfirmedGift] = useState(false);
-  const [slideExitSpeed, setSlideExitSpeed] = useState(0);
   const { enqueueSnackbar } = useSnackbar();
   const isMountedRef = useIsMountedRef();
+  const { t } = useTranslation('ConsumeGiftForm');
 
-  const { t } = useTranslation('BuildGiftForm');
-
-  const networkBlockIndexBigInt = BigInt(selectedAccount.balanceStatus.networkBlockIndex);
-  const accountBlockIndexBigInt = BigInt(selectedAccount.balanceStatus.accountBlockIndex);
-
-  const isSynced = isSyncedBuffered(networkBlockIndexBigInt, accountBlockIndexBigInt);
-
-  // TODO - consider adding minimum gift ~ 1 MOB
   // We'll use this array in prep for future patterns with multiple accounts
+  // TODO - fix the type for Account
   const mockMultipleAccounts: Array<Account> = [
     {
       b58Code: selectedAccount.account.mainAddress,
@@ -140,14 +106,7 @@ const BuildGiftForm: FC<BuildGiftFormProps> = ({
     },
   ];
 
-  const handleShowCode = () => {
-    setSlideExitSpeed(1000);
-    setShowCode(true);
-  };
-
   const handleClose = () => {
-    setSlideExitSpeed(0);
-    setShowCode(false);
     setShowModal(false);
     setIsAwaitingConformation(false);
     setConfirmation(EMPTY_CONFIRMATION);
@@ -160,28 +119,25 @@ const BuildGiftForm: FC<BuildGiftFormProps> = ({
     setSubmittingConfirmedGift(true);
     setShowModal(false);
     try {
-      if (confirmation.txProposal === null || confirmation.txProposal === undefined) {
-        throw new Error(t('confirmationNotFound'));
-      }
-
-      await submitGiftCode({
-        fromAccountId: selectedAccount.account.accountId,
+      await claimGiftCode({
+        accountId: selectedAccount.account.accountId,
         giftCodeB58: confirmation.giftCodeB58,
-        txProposal: confirmation.txProposal,
       });
+
       if (isMountedRef.current) {
         setStatus({ success: true });
         setSubmittingConfirmedGift(false);
         setIsAwaitingConformation(false);
         setConfirmation(EMPTY_CONFIRMATION);
-        enqueueSnackbar(t('giftCreated'), {
+        enqueueSnackbar(t('confirmation'), {
           variant: 'success',
         });
       }
     } catch (err) {
       if (isMountedRef.current) {
+        const santitizedError = err.message === CLAIMED_GIFT_ERROR ? t('giftClaimed') : err.message;
         setStatus({ success: false });
-        setErrors({ submit: err.message });
+        setErrors({ submit: santitizedError });
         setSubmittingConfirmedGift(false);
         setIsAwaitingConformation(false);
         setConfirmation(EMPTY_CONFIRMATION);
@@ -192,6 +148,7 @@ const BuildGiftForm: FC<BuildGiftFormProps> = ({
     }
   };
 
+  // TODO - not sure what this is about. Should delete after confirming it's useless
   // const createAccountLabel = (account: Account) => {
   //   const name =
   //     account.name && account.name.length > 0 ? `${account.name}: ` : `${t('unnamed')}: `;
@@ -211,8 +168,8 @@ const BuildGiftForm: FC<BuildGiftFormProps> = ({
   //   );
   // };
 
-  // TODO - reintroduce with multiple accounts
-  // const renderSenderPublicAddressOptions = (accounts: Account[], isSubmitting: boolean) => (
+  // TODO Reintroduce with multiple accounts
+  // const renderSenderPublicAdddressOptions = (accounts: Account[], isSubmitting: boolean) => (
   //   <Box pt={2}>
   //     <FormLabel className={classes.form} component="legend">
   //       <Typography color="primary">{t('select')}</Typography>
@@ -240,67 +197,50 @@ const BuildGiftForm: FC<BuildGiftFormProps> = ({
   //   </Box>
   // );
 
-  const validateAmount = (selectedBalance: bigint, fee: bigint) => (valueString: string) => {
-    let error;
-    const valueAsPicoMob = BigInt(valueString.replace('.', ''));
-    if (valueAsPicoMob + fee + fee > selectedBalance) {
-      // TODO - probably want to replace this before launch
-      error = t('errorFee');
-    }
-    return error;
-  };
-
-  // We'll use this to auto-select all text when focused. This is a better user
-  // experience than having to click the left-most area to start typing (else)
-  // having to spam backspace.
-  const handleSelect = (event: ChangeEvent<HTMLInputElement>) => {
-    event.target.select();
-  };
-
   return (
     <Formik
       initialValues={{
-        feeAmount: '0.010000000000', // TODO we need to pull this from constants
-        mobValue: '0', // mobs
-        pin: '',
+        giftCodeB58: '', // mobs
         senderPublicAddress: mockMultipleAccounts[0].b58Code,
         submit: null,
       }}
       validationSchema={Yup.object().shape({
-        mobValue: Yup.number()
-          .positive(t('positiveValidation'))
-          .required(t('positiveValidationRequired')),
+        giftCodeB58: Yup.string().required(t('giftB58Validation')),
       })}
       onSubmit={async (values, { setErrors, setStatus, setSubmitting, resetForm }) => {
-        // On submit, let's build the TxProposal.
-        // On success, we set the TxProposal in state.
-        // That triggers the confirmation modal.
-        // In the modal, the user has to confirm that they have written the
-        // code down (similiar to create new account modal).
-        // After they do, they can "Create Gift"
         try {
           setIsAwaitingConformation(true);
-
-          const adjustedValue = Number(values.mobValue) + 0.01;
-
-          const result = await buildGiftCode({
-            accountId: selectedAccount.account.accountId,
-            valuePmob: convertMobStringToPicoMobString(String(adjustedValue)),
-          });
+          const result = await checkGiftCodeStatus({ giftCodeB58: values.giftCodeB58 });
           if (result === null || result === undefined) {
-            throw new Error(t('errorBuild'));
+            throw new Error(t('giftB58Error'));
           }
 
-          const { feeConfirmation, giftCodeB58, totalValueConfirmation, txProposal } = result;
+          const { giftCodeStatus, giftCodeValue } = result;
 
           setConfirmation({
-            feeConfirmation,
-            giftCodeB58,
-            totalValueConfirmation,
-            txProposal,
+            giftCodeB58: values.giftCodeB58,
+            giftCodeStatus,
+            giftValue: giftCodeValue,
           });
 
-          setShowModal(true);
+          if (giftCodeStatus === 'GiftCodeAvailable') {
+            setShowModal(true);
+          } else {
+            setStatus({ success: false });
+            setSubmitting(false);
+            setIsAwaitingConformation(false);
+            if (giftCodeStatus === 'GiftCodeSubmittedPending') {
+              enqueueSnackbar(t('giftB58Error'), {
+                variant: 'warning',
+              });
+            }
+
+            if (giftCodeStatus === 'GiftCodeClaimed') {
+              enqueueSnackbar(t('giftClaimed'), {
+                variant: 'warning',
+              });
+            }
+          }
 
           if (isMountedRef.current) {
             setSubmitting(false);
@@ -316,35 +256,24 @@ const BuildGiftForm: FC<BuildGiftFormProps> = ({
         }
       }}
     >
-      {({ errors, isSubmitting, isValid, dirty, submitForm, values, setErrors, setStatus }) => {
+      {({ errors, isSubmitting, dirty, isValid, submitForm, setErrors, setStatus, values }) => {
         const selectedBalance =
           // TODO -- this is fine. we'll gut it anyway once we add multiple accounts
           // eslint-disable-next-line
           // @ts-ignore
-          BigInt(
-            mockMultipleAccounts.find((account) => account.b58Code === values.senderPublicAddress)
-              .balance
-          );
-
-        let isPinRequiredForTransaction = false;
-        if (confirmation.totalValueConfirmation) {
-          isPinRequiredForTransaction =
-            confirmation?.totalValueConfirmation + confirmation?.feeConfirmation >=
-            BigInt(pinThresholdPmob);
-        }
-
-        let remainingBalance;
+          mockMultipleAccounts.find((account) => account.b58Code === values.senderPublicAddress)
+            .balance;
+        let increasedBalance;
         let totalSent;
-        if (confirmation?.totalValueConfirmation && confirmation?.feeConfirmation) {
-          remainingBalance =
-            selectedBalance -
-            (confirmation?.totalValueConfirmation + confirmation?.feeConfirmation);
-          totalSent = confirmation?.totalValueConfirmation + confirmation?.feeConfirmation;
+
+        if (confirmation?.giftValue) {
+          increasedBalance = Number(selectedBalance) + confirmation?.giftValue - 10000000000;
+          totalSent = confirmation?.giftValue;
         }
 
         return (
           <Form>
-            {/* {renderSenderPublicAddressOptions(mockMultipleAccounts, isSubmitting)} */}
+            {/* {renderSenderPublicAdddressOptions(mockMultipleAccounts, isSubmitting)} */}
             <Box pt={4}>
               <FormLabel component="legend">
                 <Typography color="primary">{t('giftDetails')}</Typography>
@@ -352,24 +281,11 @@ const BuildGiftForm: FC<BuildGiftFormProps> = ({
               <Field
                 component={TextField}
                 fullWidth
-                label={t('mobLabel')}
+                label={t('giftCode')}
                 margin="normal"
-                name="mobValue"
-                id="mobValue"
+                name="giftCodeB58"
+                id="giftCodeB58"
                 type="text"
-                onFocus={handleSelect}
-                validate={validateAmount(
-                  selectedBalance,
-                  BigInt(values.feeAmount * 1_000_000_000_000)
-                )}
-                InputProps={{
-                  inputComponent: MOBNumberFormat,
-                  startAdornment: (
-                    <InputAdornment position="start">
-                      <MOBIcon height={20} width={20} />
-                    </InputAdornment>
-                  ),
-                }}
               />
             </Box>
             {errors.submit && (
@@ -377,17 +293,12 @@ const BuildGiftForm: FC<BuildGiftFormProps> = ({
                 <FormHelperText error>{errors.submit}</FormHelperText>
               </Box>
             )}
-            {!isSynced && (
-              <Box mt={3}>
-                <FormHelperText error>{t('errorSyncBeforeSending')}</FormHelperText>
-              </Box>
-            )}
             <SubmitButton
-              disabled={!dirty || !isSynced || !isValid || isSubmitting}
+              disabled={!dirty || !isValid || isSubmitting}
               onClick={submitForm}
               isSubmitting={isAwaitingConformation || isSubmitting}
             >
-              {isSynced ? t('createGift') : `${t('walletSyncing')}...`}
+              {t('openGift')}
             </SubmitButton>
             <Modal
               aria-labelledby="transition-modal-title"
@@ -403,15 +314,12 @@ const BuildGiftForm: FC<BuildGiftFormProps> = ({
               disableAutoFocus
               disableEnforceFocus
             >
-              <Slide in={showModal} timeout={{ enter: 0, exit: slideExitSpeed }}>
+              <Slide in={showModal} timeout={{ enter: 0, exit: 0 }}>
                 <Container className={classes.paper}>
-                  <Box py={2}>
-                    <Typography variant="h1" color="textPrimary">
-                      {t('giftConfirmation')}
-                    </Typography>
-                    {/* <Typography variant="p" color="textPrimary"> */}
-                    <Typography color="textPrimary">{t('giftConfirmationDescription')}:</Typography>
-                  </Box>
+                  <Typography color="textPrimary" variant="h1" id="transition-modal-title">
+                    {t('giftConfirmation')}
+                  </Typography>
+                  <Box py={2} />
                   <Box display="flex" justifyContent="space-between">
                     <Typography color="textPrimary">{t('accountBalance')}:</Typography>
                     <Typography color="textPrimary">
@@ -427,14 +335,12 @@ const BuildGiftForm: FC<BuildGiftFormProps> = ({
                     <Typography color="textPrimary">---</Typography>
                   </Box>
                   <Box display="flex" justifyContent="space-between">
-                    <Typography color="primary">{t('giftValue')}:</Typography>
-                    <Typography color="primary">
+                    <Typography color="textPrimary">{t('total')}:</Typography>
+                    <Typography color="textPrimary">
                       <MOBNumberFormat
                         suffix=" MOB"
                         valueUnit="pMOB"
-                        value={(
-                          confirmation?.totalValueConfirmation - confirmation?.feeConfirmation
-                        ).toString()}
+                        value={totalSent?.toString() as string}
                       />
                     </Typography>
                   </Box>
@@ -444,68 +350,36 @@ const BuildGiftForm: FC<BuildGiftFormProps> = ({
                       <MOBNumberFormat
                         suffix=" MOB"
                         valueUnit="pMOB"
-                        value={(
-                          confirmation?.feeConfirmation + confirmation?.feeConfirmation
-                        ).toString()}
+                        value={(10000000000).toString()}
                       />
                     </Typography>
                   </Box>
                   <Box display="flex" justifyContent="space-between">
-                    <Typography color="textPrimary">{t('total')}:</Typography>
-                    <Typography color="textPrimary">
-                      <MOBNumberFormat
-                        suffix=" MOB"
-                        valueUnit="pMOB"
-                        value={totalSent?.toString()}
-                      />
-                    </Typography>
-                  </Box>
-                  <Box display="flex" justifyContent="space-between">
-                    <Typography color="textPrimary">---</Typography>
-                    <Typography color="textPrimary">---</Typography>
-                  </Box>
-                  <Box display="flex" justifyContent="space-between">
-                    <Typography color="primary">{t('remaining')}:</Typography>
+                    <Typography color="primary">{t('giftValue')}:</Typography>
                     <Typography color="primary">
                       <MOBNumberFormat
                         suffix=" MOB"
                         valueUnit="pMOB"
-                        value={remainingBalance?.toString()}
+                        value={(confirmation?.giftValue - 10000000000).toString()}
+                      />
+                    </Typography>
+                  </Box>
+                  <Box display="flex" justifyContent="space-between">
+                    <Typography>---</Typography>
+                    <Typography>---</Typography>
+                  </Box>
+                  <Box display="flex" justifyContent="space-between">
+                    <Typography color="primary">{t('newBalance')}:</Typography>
+                    <Typography color="primary">
+                      <MOBNumberFormat
+                        suffix=" MOB"
+                        valueUnit="pMOB"
+                        value={increasedBalance?.toString() as string}
                       />
                     </Typography>
                   </Box>
                   <Box py={1} />
-                  {/* TODO - after multiple accounts, we should actually store these gift codes. please check jira for full explation */}
-                  <Typography variant="body2" color="textPrimary">
-                    {t('mobWillBeSent')}
-                  </Typography>
-                  <Box py={1} />
-                  {showCode ? (
-                    <AccountCard
-                      isGift
-                      account={{
-                        b58Code: confirmation?.giftCodeB58,
-                        name: t('pending'),
-                      }}
-                    />
-                  ) : (
-                    <Box display="flex" justifyContent="center" py={27}>
-                      <Button color="secondary" size="large" onClick={handleShowCode}>
-                        {t('showCode')}
-                      </Button>
-                    </Box>
-                  )}
-                  {isPinRequiredForTransaction && (
-                    <Field
-                      component={TextField}
-                      fullWidth
-                      label={t('enterPin')}
-                      margin="normal"
-                      name="pin"
-                      type="password"
-                    />
-                  )}
-                  <Box display="flex" justifyContent="space-between" style={{ padding: '1rem' }}>
+                  <Box display="flex" justifyContent="space-between">
                     <Button
                       className={classes.button}
                       color="secondary"
@@ -520,14 +394,11 @@ const BuildGiftForm: FC<BuildGiftFormProps> = ({
                     <Button
                       className={classes.button}
                       color="secondary"
-                      disabled={
-                        !showCode || (isPinRequiredForTransaction && values.pin !== existingPin)
-                      }
                       fullWidth
                       onClick={handleConfirm(setErrors, setStatus)}
                       variant="contained"
                     >
-                      {showCode ? t('confirmGift') : t('secureCode')}
+                      {t('claimGift')}
                     </Button>
                   </Box>
                 </Container>
@@ -558,5 +429,5 @@ const BuildGiftForm: FC<BuildGiftFormProps> = ({
   );
 };
 
-export default BuildGiftForm;
-export { BuildGiftForm };
+export default ConsumeGiftForm;
+export { ConsumeGiftForm };
