+++ app/pages/Gifts/ConsumeGiftPanel.view/index.ts
@@ -1,4 +1,4 @@
-import { ConsumeGiftPanel } from './ConsumeGiftPanel';
+import { ConsumeGiftPanel } from './ConsumeGiftPanel.view';
 
 export default ConsumeGiftPanel;
 export { ConsumeGiftPanel };
