+++ package.json
@@ -180,12 +180,12 @@
     "@babel/preset-react": "^7.10.4",
     "@babel/preset-typescript": "^7.10.4",
     "@babel/register": "^7.10.5",
-    "@storybook/addon-actions": "^6.2.5",
-    "@storybook/addon-essentials": "^6.2.5",
-    "@storybook/addon-links": "^6.2.5",
-    "@storybook/addons": "^6.2.5",
-    "@storybook/cli": "^6.2.5",
-    "@storybook/react": "^6.2.5",
+    "@storybook/addon-actions": "^6.2.8",
+    "@storybook/addon-essentials": "^6.2.8",
+    "@storybook/addon-links": "^6.2.8",
+    "@storybook/addons": "^6.2.8",
+    "@storybook/cli": "^6.2.8",
+    "@storybook/react": "^6.2.8",
     "@teamsupercell/typings-for-css-modules-loader": "^2.2.1",
     "@testing-library/jest-dom": "^5.11.8",
     "@testing-library/react": "^11.2.2",
